fix: update subquery used to retrieve nodes using `last` argument
There was a change in the signature of the `orderNodesBy`, to accept an
object as the second argument and this call wasn't updated.
This was causing a bug when using the `last` argument on a paginated
graphql query. The "order by" option of the sql wasn't properly set and
returned values were incorrect.